スクリプトファイルのファイル配置を修正
